gsc_function_app/requirements.txt
azure-functions
azure-storage-blob
beautifulsoup4
psycopg[binary]
